+++ CHANGELOG.md
@@ -2,10 +2,15 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
-## [36.4.0] 2024-08-02
+## [36.5.0] 2024-08-02
 ### Added
 - Added support of disabledDisplay in R5 Questionnaire.
  
+## [36.4.0] 2024-07-23
+### Added
+- Support for rendering-xhtml on item.text.
+- Added new template option allowHTML. allowHTMLInInstructions is deprecated.
+
 ## [36.3.3] 2024-07-22
 ### Added
 - Copy questionnaire-ordinalValue (STU3), ordinalValue (R4) and